+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - enhancements
+    + \#52830: Add cancel button to nested forms
 - bugs
     + \#52680: Add spacing between label and helper text on checkbox fields
     + \#51982: Vertical checkboxes should be vertical